+++ api/src/main/java/com/illtamer/infinite/bot/api/event/notice/FriendMessageRecallEvent.java
@@ -30,13 +30,13 @@ public class FriendMessageRecallEvent extends NoticeEvent {
      * 被撤回的消息 ID
      * */
     @SerializedName("message_id")
-    private Long messageId;
+    private Integer messageId;
 
     /**
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(String message) {
+    public Integer sendMessage(String message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
 
@@ -44,7 +44,7 @@ public class FriendMessageRecallEvent extends NoticeEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(Message message) {
+    public Integer sendMessage(Message message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
