+++ api/src/main/java/com/illtamer/infinite/bot/api/event/notice/FriendNoticeEvent.java
@@ -30,7 +30,7 @@ public class FriendNoticeEvent extends NoticeEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(String message) {
+    public Integer sendMessage(String message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
 
@@ -38,7 +38,7 @@ public class FriendNoticeEvent extends NoticeEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(Message message) {
+    public Integer sendMessage(Message message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
