+++ api/src/main/java/com/illtamer/infinite/bot/api/event/notice/NudgeEvent.java
@@ -66,7 +66,7 @@ public class NudgeEvent extends NoticeEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(String message) {
+    public Integer sendMessage(String message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
 
@@ -74,7 +74,7 @@ public class NudgeEvent extends NoticeEvent {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(Message message) {
+    public Integer sendMessage(Message message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
