+++ api/src/main/java/com/illtamer/infinite/bot/api/event/request/FriendRequestEvent.java
@@ -57,7 +57,7 @@ public class FriendRequestEvent extends RequestEvent implements QuickAction {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(String message) {
+    public Integer sendMessage(String message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
 
@@ -65,7 +65,7 @@ public class FriendRequestEvent extends RequestEvent implements QuickAction {
      * 向该消息发送者发送消息
      * @return 消息 ID
      * */
-    public Double sendMessage(Message message) {
+    public Integer sendMessage(Message message) {
         return OpenAPIHandling.sendMessage(message, userId);
     }
